fixed warning from GCC;

+++ src/support.c
@@ -20,13 +20,13 @@ const char * TymFunctionCommandMapping[] =
 
 const char *
 tym_functions(void) {
-  unsigned string_length = 0;
+  size_t string_length = 0;
   for (unsigned i = 0; i < TYM_NO_FUNCTION; ++i) {
      string_length += strlen(TymFunctionCommandMapping[i]) + 2;
   }
   char * result = malloc(string_length + 1);
   const char * sep = ", ";
-  for (unsigned offset = 0, i = 0; i < TYM_NO_FUNCTION; ++i) {
+  for (size_t offset = 0, i = 0; i < TYM_NO_FUNCTION; ++i) {
      strcpy(result + offset, TymFunctionCommandMapping[i]);
      offset += strlen(TymFunctionCommandMapping[i]);
      if (TYM_NO_FUNCTION - 1 != i) {
